Proper usage of pixel pack buffer
The pixel pack buffer was ignored, but should be
used whenever it is set when doing a glReadPixels.
When this is the case, pixels becomes an offset.
From the GLES 3.0 spec, section 4.3.1, subsection
"Placement in Pixel Pack Buffer or Client Memory":

"If a pixel pack buffer is bound, data is an offset
into the pixel pack buffer and the pixels are
packed into the buffer relative to this offset;
otherwise, data is a pointer to a block of client
memory and the pixels are packed into the client
memory relative to the pointer"

WebGL conformance tests passed.

Change-Id: I7539f1ce705cfd2072ed679815676c0b18887272
Reviewed-on: https://swiftshader-review.googlesource.com/3604
Tested-by: Alexis Hétu <sugoi@google.com>
Reviewed-by: Nicolas Capens <capn@google.com>