Update patch set 2
Patch Set 2:

I'm not sure I fully agree with the motivation for TConstParameter. It's part of a debated larger change to save a few kB at the cost of adding complication: https://chromium-review.googlesource.com/281411 I'll try to discuss this with Dmitry and check if there's a more elegant solution. In the meantime feel free to split this patch to implement "length()", if that's not too hard.

Patch-set: 2
Label: Code-Review=0