Update patch set 3
Patch Set 3: Code-Review-1

(7 comments)

I just noticed in the spec that "If the logical operation is enabled for color values, it is as if blending were disabled, regardless of the value of BLEND."

This means it might be more logical (no pun intended) to perform the logic op within the same function as alpha blending, and just treat them as additional blend modes. But let's leave that as a separate refactoring patch to keep this patch simple, and make a note of it in the Buganizer issue.

Patch-set: 3
Label: Code-Review=-1