Update patch set 1
Patch Set 1: Code-Review-1

It's currently not intended to be changed through the build configuration. It's a hardcoded constant, and this CL merely gives it a symbolic name in Config.hpp to keep it grouped with other constants.

I realize it's not ideal, and I could easily be persuaded to make it a macro that can be overridden in the build config...

Patch-set: 1
Label: Code-Review=-1