Update patch set 3
Patch Set 3: Verified+1

Patch set 3 is a lot cleaner. There's still more state that needs to be saved/restored, and there's probably an overall better way to implement this, but at least it should be an incremental improvement.

I verified that the boot logo still renders fine in the emulator. Please test the camera app on JB-MR1 again.

Patch-set: 3
Label: Verified=+1