Subzero: Remove an unwanted assert.
At the end, the ELF emitter code marks all the external symbols that are undefined in this object file, but it filters intrinsics out of this set.  Previously, it was asserting on a "bad intrinsic", meaning a function whose name starts with "llvm." but doesn't match a known intrinsic.

The assert doesn't really make sense here.  Either it should be more strongly validated early on in the translation, or it should just pass through (and the link will probably fail, or maybe not).

BUG= https://code.google.com/p/nativeclient/issues/detail?id=4319
R=kschimpf@google.com

Review URL: https://codereview.chromium.org/1343283002 .