Fix pnacl-sz to return with staus 0 in report_fatal_error.
Fixes pnacl-sz to return with exit status 0 in report_fatal_error, if
command line flag --exit-status is specified. The importance of this
is that it allows afl-fuzz to not report the mutation as a crash.

In addition, afl-fuzz doesn't record crash paths in its search
history.  By returning success, afl-fuzz can continue to apply
additional mutations to the bad input. This allows afl-fuzz to add
errors that require multiple changes to occur on the input.

BUG=None
R=stichnot@chromium.org

Review URL: https://codereview.chromium.org/1382653002 .