Subzero: Don't "and" i1 values with 1.
In x86 lowering, i1 values are held in i8 register and memory slots.  We were conservatively "and"ing  them with 1 before zero-extending them for some lowering operations, but this "and" with 1 is unnecessary and just clutters the code.

We continue the invariant that all i1-produced values in an i8 slot are either 0 or 1.

BUG= https://code.google.com/p/nativeclient/issues/detail?id=4095
R=jpp@chromium.org

Review URL: https://codereview.chromium.org/1394413002 .