Update patch set 1
Patch Set 1: Code-Review-1

VertexProgram::readConstant() prevents reading outside of the 256 vertex constants for dynamic indexing, and we'll have to continue to ensure that (0, 0, 0, 0) is read in that case. Anthony was also looking into this as part of the robustness extension.

Much of this will change when uniform buffers are properly supported (which can be way bigger), so maybe we should discuss the implications at our next meeting.

Patch-set: 1
Reviewer: Nicolas Capens <5005@aea13c4a-0b89-3eca-aee9-e193b1b77aa4>
Label: Code-Review=-1