Implements simple returns and call args for Mips. This patch is essentially the same as for ARM https://codereview.chromium.org/1127963004
I have incorporated the new 64 bit register work which was not available
at the time of this earlier patch.

The MIPS O32 Abi is not perfect on this patch but I am more or less following
the development of the ARM patches and those were preliminary at this
stage too. I will make corrections in a later patch when I incorporate
more of the ARM patches.

BUG= https://code.google.com/p/nativeclient/issues/detail?id=4167
R=stichnot@chromium.org

Review URL: https://codereview.chromium.org/1416493002 .