Update patch set 3
Patch Set 3:

(1 comment)

Sorry, for the increase in patch size, I was trying to preempt your review, but didn't make it in time :)

So I added a little bit more meat around this cl to make the goal clearer. The extra changes don't allow more texture functions to be used yet, but the changes now extend all the way down to the program function stubs.

Patch-set: 3
Label: Code-Review=0